requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
praw==7.7.1
numpy==1.24.3
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.3.0
python-dotenv==1.0.0
tqdm==4.65.0
selenium==4.18.1 